requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# OpenAI and AI
openai==1.6.1
tiktoken==0.5.1

# Database
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0  # For PostgreSQL
aiosqlite==0.19.0  # For SQLite

# Data processing and validation
pydantic==2.5.0
pandas==2.1.4
numpy==1.24.3

# Web and networking
websockets==12.0
httpx==0.25.2
aiofiles==23.2.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment and configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Monitoring and logging
loguru==0.7.2
prometheus-client==0.19.0

# Analytics and metrics
plotly==5.17.0
scipy==1.11.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.12.0
isort==5.13.1
flake8==6.1.0

# Optional: Redis for caching
redis==5.0.1
aioredis==2.0.1